third_party/blink/web_tests/fast/events/scroll-event-handler-reused-window-expected.txt

Test tracking event handlers with a reused DOMWindow.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS internals.scrollEventHandlerCount(newWindow.document) is 1
